requirements.txt
fastapi
uvicorn
jinja2
python-multipart
pillow
bcrypt
python-dotenv
alembic
captcha
psycopg2
PyJWT